move BackupContext interfaces and it's abstract extend classes to service module
